requirements.txt
# KIS US Auto Trading System Requirements

# Core dependencies
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4

# Technical Analysis
ta==0.11.0

# Async WebSocket support
websockets==14.1

# Date/Time handling
python-dateutil==2.9.0.post0
pytz==2024.1